+++ web/src/main/java/project/pa165/musiclibrary/web/rest/SongController.java
@@ -1,13 +1,16 @@
 package project.pa165.musiclibrary.web.rest;
 
 import org.springframework.http.HttpStatus;
+import org.springframework.validation.Errors;
 import org.springframework.web.bind.annotation.*;
 import project.pa165.musiclibrary.dto.SongDto;
+import project.pa165.musiclibrary.exception.SongBadRequestException;
 import project.pa165.musiclibrary.exception.SongNotFoundException;
 import project.pa165.musiclibrary.services.SongService;
 import project.pa165.musiclibrary.util.ErrorInfo;
 
 import javax.inject.Inject;
+import javax.validation.Valid;
 import java.util.List;
 
 /**
@@ -29,8 +32,8 @@ public class SongController {
     }
 
     @RequestMapping(method = RequestMethod.GET)
-    public void getAllSongs() {
-        songService.getAllSongs();
+    public List<SongDto> getAllSongs() {
+        return songService.getAllSongs();
     }
 
     @RequestMapping(value = "/search", method = RequestMethod.GET)
@@ -47,17 +50,22 @@ public class SongController {
 
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
     public void deleteSong(@PathVariable("id") Long id) {
+        if (getSongService().findSong(id) == null) throw new SongNotFoundException(id.toString());
         songService.deleteSong(id);
     }
 
     @RequestMapping(method = RequestMethod.POST)
-    public void createSong(@RequestBody SongDto songDto) {
-        songService.createSong(songDto);
+    public void createSong(@Valid @RequestBody SongDto song, Errors errors) {
+        if (song == null || errors.hasErrors())
+            throw new SongBadRequestException("Failed to map JSON to SongDto");
+        songService.createSong(song);
     }
 
     @RequestMapping(method = RequestMethod.PUT)
-    public void updateSong(@RequestBody SongDto songDto) {
-        songService.updateSong(songDto);
+    public void updateSong(@Valid @RequestBody SongDto song, Errors errors) {
+        if (song == null || errors.hasErrors())
+            throw new SongBadRequestException("Failed to map JSON to SongDto");
+        songService.updateSong(song);
     }
 
     @ResponseStatus(HttpStatus.NOT_FOUND)
@@ -66,5 +74,13 @@ public class SongController {
     @ResponseBody
     ErrorInfo handleSongNotFoundException() {
         return new ErrorInfo(404, "Song could not be found");
+    }
+
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    @ExceptionHandler(SongBadRequestException.class)
+    public
+    @ResponseBody
+    ErrorInfo handleSongBadRequestException() {
+        return new ErrorInfo(400, "Song\'s method with bad request");
     }
 }
